requirements.txt
asgiref==3.5.2
autopep8==2.0.1
beautifulsoup4==4.11.2
black==22.12.0
click==8.1.3
Django==4.1.2
django-appconf==1.0.5
django-extensions==3.2.1
django-imagekit==4.1.0
django-livereload-server==0.4
django-resized==1.0.2
djangorestframework==3.14.0
geographiclib==1.52
geopy==2.2.0
livereload==2.6.3
mypy-extensions==0.4.3
pathspec==0.10.3
pilkit==2.0
Pillow==9.2.0
platformdirs==2.6.2
psycopg2-binary==2.9.4
pycodestyle==2.10.0
pytz==2022.7.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
